+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.0.2] - 2025-03-10
+### Added
+- include/exlude lists
+- Bugs, TODO
+
+## Changed
+- use dbMeshKey and dbNodeKey
+
 ## [0.0.1] - 2025-03-10
 ### Added
 - inital draft